Plots/Tables: Migrate common code to artefact_util.py
