+++ varats/varats/tables/szz_quality_metrics_table.py
@@ -17,7 +17,6 @@ from varats.table.tables import (
     TableGenerator,
     REQUIRE_MULTI_CASE_STUDY,
     OPTIONAL_REPORT_TYPE,
-    OPTIONAL_TABLE_FORMAT,
 )
 
 
@@ -78,9 +77,7 @@ class BugOverviewTable(Table):
 class BugOverviewTableGenerator(
     TableGenerator,
     generator_name="szz-quality-metrics-table",
-    options=[
-        REQUIRE_MULTI_CASE_STUDY, OPTIONAL_REPORT_TYPE, OPTIONAL_TABLE_FORMAT
-    ]
+    options=[REQUIRE_MULTI_CASE_STUDY, OPTIONAL_REPORT_TYPE]
 ):
     """Generates a szz-quality-metrics table for the selected case study."""
